Refactor to fix namespace errors
